+++ src/main/java/springschool/ranking/AppConfig.java
@@ -1,13 +1,7 @@
 package springschool.ranking;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import springschool.ranking.rank.RankPolicy;
-import springschool.ranking.rank.ScoreRankPolicy;
-import springschool.ranking.student.MemoryStudentRepository;
-import springschool.ranking.student.StudentRepository;
 
 @Configuration
 @ComponentScan
