--- src/main/java/springschool/ranking/rank/RankService.java
@@ -1,7 +1,0 @@
-package springschool.ranking.rank;
-
-import springschool.ranking.student.Student;
-
-public interface RankService {
-    public Rank createRank(Long id);
-}
