+++ src/main/java/springschool/ranking/rank/service/GradeRankPolicy.java
@@ -1,22 +1,22 @@
-package springschool.ranking.rank;
+package springschool.ranking.rank.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-import springschool.ranking.student.Student;
-import springschool.ranking.student.StudentRepository;
+import springschool.ranking.rank.Policy;
+import springschool.ranking.student.domain.Grade;
+import springschool.ranking.student.domain.Student;
+import springschool.ranking.student.repository.StudentRepository;
 
 import java.util.*;
 
 //@Component
-public class ScoreRankPolicy implements RankPolicy {
-
+public class GradeRankPolicy implements RankPolicy{
     private final StudentRepository studentRepository;
-    private static Map<Long, Integer> store = new HashMap<>();
-    private static List<Map.Entry<Long, Integer>> list = new ArrayList<>();
+    private static Map<Long, Grade> store = new HashMap<>();
+    private static List<Map.Entry<Long, Grade>> list = new ArrayList<>();
 
 
     @Autowired
-    public ScoreRankPolicy(StudentRepository studentRepository) {
+    public GradeRankPolicy(StudentRepository studentRepository) {
         this.studentRepository = studentRepository;
     }
 
@@ -27,12 +27,12 @@ public class ScoreRankPolicy implements RankPolicy {
         // Repository로부터 받은 리스트를 store에 담기
         List<Student> tmp = this.studentRepository.findAll();
         for (Student student : tmp) {
-            store.put(student.getId(), student.getScore());
+            store.put(student.getId(), student.getGrade());
         }
 
 
         // store를 ArrayList로 변환한 다음에 Collections.sort() 정렬
-        Set<Map.Entry<Long, Integer>> set = store.entrySet();
+        Set<Map.Entry<Long, Grade>> set = store.entrySet();
         list = new ArrayList<>(set); // ArrayList(Collection c)
 
         // list를 생성하자마자 정렬시켜준다.
@@ -43,7 +43,7 @@ public class ScoreRankPolicy implements RankPolicy {
     public void sortRank() {
 
         // static void sort(List list, Comparator c)
-        Collections.sort(list, new ScoreComparator());
+        Collections.sort(list, new GradeRankPolicy.GradeComparator());
     }
 
     @Override
@@ -51,8 +51,8 @@ public class ScoreRankPolicy implements RankPolicy {
 
         setList();
 
-        int i = 0;
-        for (Map.Entry<Long, Integer> entry : list) {
+        int i=0;
+        for (Map.Entry<Long, Grade> entry : list) {
             if (entry.getKey() == student.getId()) {
                 break;
             }
@@ -66,37 +66,40 @@ public class ScoreRankPolicy implements RankPolicy {
     public void printRankList() {
         setList();
 
-        Iterator<Map.Entry<Long, Integer>> it = list.iterator();
+        Iterator<Map.Entry<Long, Grade>> it = list.iterator();
 
-        System.out.println("= 점수의 크기가 큰 순서로 정렬 =");
+        System.out.println("= 등급의 크기가 작은 순서로 정렬 =");
         while (it.hasNext()) {
-            Map.Entry<Long, Integer> entry = it.next();
+            Map.Entry<Long, Grade> entry = it.next();
             String name = studentRepository.findById(entry.getKey()).getName();
-            int score = entry.getValue().intValue();
-            System.out.println(entry.getKey() + " : " + name + " : " + score);
+            Grade grade = entry.getValue();
+            System.out.println(entry.getKey() + " : " + name + " : " + grade);
         }
         System.out.println();
     }
 
     @Override
     public Policy getPolicy() {
-        return Policy.SCORE;
+        return Policy.GRADE;
     }
 
     /**
-     * Map 자료형을 점수가 큰 사람부터 내림차순을 하기 위한 Comparator 구현
+     * Map 자료형을 등급이 작은 사람부터 오름차순을 하기 위한 Comparator 구현
      */
-    static class ScoreComparator implements Comparator {
+    static class GradeComparator implements Comparator {
         @Override
         public int compare(Object o1, Object o2) {
             if (o1 instanceof Map.Entry<?, ?> && o2 instanceof Map.Entry<?, ?>) {
                 Map.Entry e1 = (Map.Entry) o1;
                 Map.Entry e2 = (Map.Entry) o2;
 
-                int v1 = ((Integer) e1.getValue()).intValue();
-                int v2 = ((Integer) e2.getValue()).intValue();
+                Grade g1 = (Grade) e1.getValue();
+                Grade g2 = (Grade) e2.getValue();
 
-                return v2 - v1;
+                int v1 = g1.getValue();
+                int v2 = g2.getValue();
+
+                return v1 - v2;
             }
             return -1;
         }
