+++ src/main/java/springschool/ranking/rank/service/RankService.java
@@ -1,0 +1,7 @@
+package springschool.ranking.rank.service;
+
+import springschool.ranking.rank.domain.Rank;
+
+public interface RankService {
+    public Rank createRank(Long id);
+}
