+++ src/main/java/springschool/ranking/rank/service/RankServiceImpl.java
@@ -1,9 +1,10 @@
-package springschool.ranking.rank;
+package springschool.ranking.rank.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import springschool.ranking.student.Student;
-import springschool.ranking.student.StudentRepository;
+import springschool.ranking.rank.domain.Rank;
+import springschool.ranking.student.domain.Student;
+import springschool.ranking.student.repository.StudentRepository;
 
 @Service
 public class RankServiceImpl implements RankService {
