+++ src/main/java/springschool/ranking/student/domain/Grade.java
@@ -1,4 +1,4 @@
-package springschool.ranking.student;
+package springschool.ranking.student.domain;
 
 public enum Grade {
     ONE(1), TWO(2), THREE(3), FOUR(4), FIVE(5), SIX(6), SEVEN(7), EIGHT(8), NINE(9);
