+++ src/main/java/springschool/ranking/student/repository/StudentRepository.java
@@ -1,4 +1,6 @@
-package springschool.ranking.student;
+package springschool.ranking.student.repository;
+
+import springschool.ranking.student.domain.Student;
 
 import java.util.List;
