+++ src/main/java/springschool/ranking/student/service/StudentService.java
@@ -1,4 +1,6 @@
-package springschool.ranking.student;
+package springschool.ranking.student.service;
+
+import springschool.ranking.student.domain.Student;
 
 public interface StudentService {
     void register(Student student);
