+++ src/main/java/springschool/ranking/student/service/StudentServiceImpl.java
@@ -1,7 +1,9 @@
-package springschool.ranking.student;
+package springschool.ranking.student.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import springschool.ranking.student.domain.Student;
+import springschool.ranking.student.repository.StudentRepository;
 
 @Service
 public class StudentServiceImpl implements StudentService{
